Add eslint and clean most (easy) errors
eslint was added to help ensure consistency in the codebase and help
keep readability and maintainability. Applied airbnb for the base rules
and added a few overrides for what makes sense for us in Node.
gitattributes was added to help with eslinting consistencies across
Windows and Linux platforms.

Also added Jest to the package list to prep for unit testing. Removed
package-lock because we are sticking with yarn instead of npm.